+++ stableset/test/test_ordered_set_2.py
@@ -1,387 +1,586 @@
-import collections
-import itertools as it
-import operator
+# The following tests are based on
+# https://github.com/simonpercivall/orderedset/blob/master/tests/test_orderedset.py
+# Thus shall be under the license:
+# https://github.com/simonpercivall/orderedset/blob/master/LICENSE
+
+import copy
+import gc
 import pickle
-import random
-import sys
+import weakref
 
 import pytest
 
-from stableset import OrderedSet
+from stableset.test import (
+    all_sets,
+    ordered_sets,
+    stable_and_ordered_sets,
+    stableeq_and_ordered_sets,
+    stableeq_sets,
+)
+from stableset.test.pytest_util import (
+    assertEqual,
+    assertFalse,
+    assertGreater,
+    assertGreaterEqual,
+    assertIsNot,
+    assertLess,
+    assertLessEqual,
+    assertNotEqual,
+    assertNotIn,
+    assertTrue,
+)
 
+datasets = [list(range(10))]
 
-def test_pickle():
-    set1 = OrderedSet("abracadabra")
-    roundtrip = pickle.loads(pickle.dumps(set1))
-    assert roundtrip == set1
 
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_add_new(set_t, lst: list):
+    oset = set_t(lst)
+    lst = copy.copy(lst)
 
-def test_empty_pickle():
-    empty_oset = OrderedSet()
-    empty_roundtrip = pickle.loads(pickle.dumps(empty_oset))
-    assert empty_roundtrip == empty_oset
+    item = 10
+    lst.append(item)
+    oset.add(item)
 
+    assertEqual(list(oset), lst)
 
-def test_order():
-    set1 = OrderedSet("abracadabra")
-    assert len(set1) == 5
-    assert set1 == OrderedSet(["a", "b", "r", "c", "d"])
-    assert list(reversed(set1)) == ["d", "c", "r", "b", "a"]
 
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_add_existing(set_t, lst: list):
+    oset = set_t(lst)
+    lst = copy.copy(lst)
 
-def test_binary_operations():
-    set1 = OrderedSet("abracadabra")
-    set2 = OrderedSet("simsalabim")
-    assert set1 != set2
+    oset.add(1)
+    oset.add(3)
+    assertEqual(list(oset), lst)
 
-    assert set1 & set2 == OrderedSet(["a", "b"])
-    assert set1 | set2 == OrderedSet(["a", "b", "r", "c", "d", "s", "i", "m", "l"])
-    assert set1 - set2 == OrderedSet(["r", "c", "d"])
 
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_discard(set_t, lst: list):
+    oset = set_t([1, 2, 3])
 
-def test_indexing():
-    set1 = OrderedSet("abracadabra")
-    assert set1[:] == set1
-    assert set1.copy() == set1
-    assert set1 is set1
-    assert set1[:] is not set1
-    assert set1.copy() is not set1
+    oset.discard(1)
+    assertNotIn(1, oset)
 
-    assert set1[[1, 2]] == OrderedSet(["b", "r"])
-    assert set1[1:3] == OrderedSet(["b", "r"])
-    assert set1.index("b") == 1
-    assert set1.index(["b", "r"]) == [1, 2]
-    with pytest.raises(KeyError):
-        set1.index("br")
+    oset.discard(4)
 
 
-class FancyIndexTester:
-    """
-    Make sure we can index by a NumPy ndarray, without having to import
-    NumPy.
-    """
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_pop(set_t):
+    oset = set_t([1, 2, 3])
 
-    def __init__(self, indices):
-        self.indices = indices
+    v = oset.pop()
+    assertEqual(v, 3)
+    assertNotIn(v, oset)
 
-    def __iter__(self):
-        return iter(self.indices)
+    v = oset.popitem(last=False)
+    assertEqual(v, 1)
+    assertNotIn(v, oset)
 
-    def __index__(self):
-        raise TypeError("NumPy arrays have weird __index__ methods")
 
-    def __eq__(self, other):
-        # Emulate NumPy being fussy about the == operator
-        raise TypeError
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_remove(set_t, lst: list):
+    oset = set_t(lst)
+    lst = copy.copy(lst)
 
+    oset.remove(3)
+    lst.remove(3)
 
-def test_fancy_index_class():
-    set1 = OrderedSet("abracadabra")
-    indexer = FancyIndexTester([1, 0, 4, 3, 0, 2])
-    assert "".join(set1[indexer]) == "badcar"
+    assertEqual(list(oset), lst)
 
 
-def test_pandas_compat():
-    set1 = OrderedSet("abracadabra")
-    assert set1.get_loc("b") == 1
-    assert set1.get_indexer(["b", "r"]) == [1, 2]
+@pytest.mark.parametrize("set_t", all_sets)
+def test_clear(set_t):
+    val = frozenset([1])
 
+    oset = set_t()
+    ws = weakref.WeakKeyDictionary()
 
-def test_tuples():
-    set1 = OrderedSet()
-    tup = ("tuple", 1)
-    set1.add(tup)
-    assert set1.index(tup) == 0
-    assert set1[0] == tup
+    oset.add(val)
+    ws[val] = 1
+    oset.clear()
 
+    assertEqual(list(oset), [])
 
-def test_remove():
-    set1 = OrderedSet("abracadabra")
+    del val
+    gc.collect()
+    assertEqual(list(ws), [])
 
-    set1.remove("a")
-    set1.remove("b")
 
-    assert set1 == OrderedSet("rcd")
-    assert set1[0] == "r"
-    assert set1[1] == "c"
-    assert set1[2] == "d"
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_copy(set_t, lst: list):
+    oset1 = set_t(lst)
+    oset2 = oset1.copy()
 
-    assert set1.index("r") == 0
-    assert set1.index("c") == 1
-    assert set1.index("d") == 2
+    assertIsNot(oset1, oset2)
+    assertEqual(oset1, oset2)
 
-    assert "a" not in set1
-    assert "b" not in set1
-    assert "r" in set1
+    oset1.clear()
+    assertNotEqual(oset1, oset2)
 
-    # Make sure we can .discard() something that's already gone, plus
-    # something that was never there
-    set1.discard("a")
-    set1.discard("a")
 
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_reduce(set_t, lst: list):
+    oset = set_t(lst)
+    oset2 = copy.copy(oset)
+    assertEqual(oset, oset2)
 
-def test_remove_error():
-    # If we .remove() an element that's not there, we get a KeyError
-    set1 = OrderedSet("abracadabra")
-    with pytest.raises(KeyError):
-        set1.remove("z")
+    oset3 = pickle.loads(pickle.dumps(oset))
+    assertEqual(oset, oset3)
 
+    oset.add(-1)
+    assertNotEqual(oset, oset2)
 
-def test_clear():
-    set1 = OrderedSet("abracadabra")
-    set1.clear()
 
-    assert len(set1) == 0
-    assert set1 == OrderedSet()
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_difference_and_update(set_t, lst: list):
+    list1 = [1, 2, 3]
+    list2 = [3, 4, 5]
+    oset1 = set_t(list1)
+    oset2 = set_t(list2)
 
+    # right hand
+    oset3 = oset2 - list1
+    assertEqual(oset3, set_t([4, 5]))
+    assertEqual(oset2.difference(oset1), oset3)
 
-def test_update():
-    set1 = OrderedSet("abcd")
-    result = set1.update("efgh")
+    oset3 = list2 - oset1
+    assertEqual(oset3, set_t([4, 5]))
+    assertEqual(oset2.difference(oset1), oset3)
 
-    assert result == 7
-    assert len(set1) == 8
-    assert "".join(set1) == "abcdefgh"
+    oset3 = oset1 - oset2
+    assertEqual(oset3, set_t([1, 2]))
+    assertEqual(oset1.difference(oset2), oset3)
 
-    set2 = OrderedSet("abcd")
-    result = set2.update("cdef")
-    assert result == 5
-    assert len(set2) == 6
-    assert "".join(set2) == "abcdef"
+    # left hand
+    oset3 = oset1 - list2
+    assertEqual(oset3, set_t([1, 2]))
+    assertEqual(oset1.difference(oset2), oset3)
 
+    oset3 = list1 - oset2
+    assertEqual(oset3, set_t([1, 2]))
+    assertEqual(oset1.difference(oset2), oset3)
 
-def test_pop():
-    set1 = OrderedSet("ab")
-    elem = set1.pop()
+    oset4 = oset1.copy()
+    oset4 -= oset2
+    assertEqual(oset4, oset3)
 
-    assert elem == "b"
-    elem = set1.pop()
+    oset5 = oset1.copy()
+    oset5.difference_update(oset2)
+    assertEqual(oset5, oset3)
 
-    assert elem == "a"
 
-    pytest.raises(KeyError, set1.pop)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_intersection_and_update(set_t):
+    oset1 = set_t([1, 2, 3])
+    oset2 = set_t([3, 4, 5])
 
+    oset3 = oset1 & oset2
+    assertEqual(oset3, set_t([3]))
 
-def test_getitem_type_error():
-    set1 = OrderedSet("ab")
-    with pytest.raises(TypeError):
-        set1["a"]
+    oset4 = oset1.copy()
+    oset4 &= oset2
 
+    assertEqual(oset4, oset3)
 
-def test_update_value_error():
-    set1 = OrderedSet("ab")
-    with pytest.raises(ValueError):
-        # noinspection PyTypeChecker
-        set1.update(3)
+    oset5 = oset1.copy()
+    oset5.intersection_update(oset2)
+    assertEqual(oset5, oset3)
 
 
-def test_empty_repr():
-    set1 = OrderedSet()
-    assert repr(set1) == "OrderedSet()"
+@pytest.mark.parametrize("set_t", all_sets)
+def test_issubset(set_t):
+    oset1 = set_t([1, 2, 3])
+    oset2 = set_t([1, 2])
 
+    assertTrue(oset2 < oset1)
+    assertTrue(oset2.issubset(oset1))
 
-def test_eq_wrong_type():
-    set1 = OrderedSet()
-    assert set1 != 2
+    oset2 = set_t([1, 2, 3])
+    assertTrue(oset2 <= oset1)
+    assertTrue(oset1 <= oset2)
+    assertTrue(oset2.issubset(oset1))
 
+    oset2 = set_t([1, 2, 3, 4])
+    assertFalse(oset2 < oset1)
+    assertFalse(oset2.issubset(oset1))
+    assertTrue(oset1 < oset2)
 
-def test_ordered_equality():
-    # Ordered set checks order against sequences.
-    assert OrderedSet([1, 2]) == OrderedSet([1, 2])
-    assert OrderedSet([1, 2]) == [1, 2]
-    assert OrderedSet([1, 2]) == (1, 2)
-    assert OrderedSet([1, 2]) == collections.deque([1, 2])
+    # issubset compares unordered for all sets
+    oset2 = set_t([4, 3, 2, 1])
+    assertTrue(oset1 < oset2)
 
 
-def test_ordered_inequality():
-    # Ordered set checks order against sequences.
-    assert OrderedSet([1, 2]) != OrderedSet([2, 1])
+@pytest.mark.parametrize("set_t", all_sets)
+def test_issuperset(set_t):
+    oset1 = set_t([1, 2, 3])
+    oset2 = set_t([1, 2])
 
-    assert OrderedSet([1, 2]) != [2, 1]
-    assert OrderedSet([1, 2]) != [2, 1, 1]
+    assertTrue(oset1 > oset2)
+    assertTrue(oset1.issuperset(oset2))
 
-    assert OrderedSet([1, 2]) != (2, 1)
-    assert OrderedSet([1, 2]) != (2, 1, 1)
+    oset2 = set_t([1, 2, 3])
+    assertTrue(oset1 >= oset2)
+    assertTrue(oset2 >= oset1)
+    assertTrue(oset1.issubset(oset2))
 
-    # Note: in Python 2.7 deque does not inherit from Sequence, but __eq__
-    # contains an explicit check for this case for python 2/3 compatibility.
-    assert OrderedSet([1, 2]) != collections.deque([2, 1])
-    assert OrderedSet([1, 2]) != collections.deque([2, 2, 1])
+    oset2 = set_t([1, 2, 3, 4])
+    assertFalse(oset1 > oset2)
+    assertFalse(oset1.issuperset(oset2))
+    assertTrue(oset2 > oset1)
 
+    # issubset compares underordered for all sets
+    oset2 = set_t([4, 3, 2, 1])
+    assertTrue(oset2 > oset1)
 
-def test_comparisons():
-    # Comparison operators on sets actually test for subset and superset.
-    assert OrderedSet([1, 2]) < OrderedSet([1, 2, 3])
-    assert OrderedSet([1, 2]) > OrderedSet([1])
 
-    # MutableSet subclasses aren't comparable to set on 3.3.
-    if tuple(sys.version_info) >= (3, 4):
-        assert OrderedSet([1, 2]) > {1}
+@pytest.mark.parametrize("set_t", all_sets)
+def test_symmetric_difference_and_update(set_t):
+    oset1 = set_t([1, 2, 3])
+    oset2 = set_t([2, 3, 4])
 
+    oset3 = oset1 ^ oset2
+    assertEqual(oset3, set_t([1, 4]))
 
-def test_unordered_equality():
-    # Unordered set checks order against non-sequences.
-    assert OrderedSet([1, 2]) == {1, 2}
-    assert OrderedSet([1, 2]) == frozenset([2, 1])
+    oset4 = oset1.copy()
+    assertEqual(oset4.symmetric_difference(oset2), oset3)
 
-    assert OrderedSet([1, 2]) == {1: "a", 2: "b"}
-    assert OrderedSet([1, 2]) == {1: 1, 2: 2}.keys()
-    assert OrderedSet([1, 2]) == {1: 1, 2: 2}.values()
+    oset4 ^= oset2
+    assertEqual(oset4, oset3)
 
-    # Corner case: OrderedDict is not a Sequence, so we don't check for order,
-    # even though it does have the concept of order.
-    assert OrderedSet([1, 2]) == collections.OrderedDict([(2, 2), (1, 1)])
+    oset5 = oset1.copy()
+    oset5.symmetric_difference_update(oset2)
+    assertEqual(oset5, oset3)
 
-    # Corner case: We have to treat iterators as unordered because there
-    # is nothing to distinguish an ordered and unordered iterator
-    assert OrderedSet([1, 2]) == iter([1, 2])
-    assert OrderedSet([1, 2]) == iter([2, 1])
-    assert OrderedSet([1, 2]) == iter([2, 1, 1])
 
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stableeq_and_ordered_sets)
+def test_union_and_update(set_t, lst: list):
+    oset = set_t(lst)
+    lst = copy.copy(lst)
 
-def test_unordered_inequality():
-    assert OrderedSet([1, 2]) != set([])
-    assert OrderedSet([1, 2]) != frozenset([2, 1, 3])
+    oset2 = oset | [3, 9, 27]
+    assertEqual(oset2, lst + [27])
 
-    assert OrderedSet([1, 2]) != {2: "b"}
-    assert OrderedSet([1, 2]) != {1: 1, 4: 2}.keys()
-    assert OrderedSet([1, 2]) != {1: 1, 2: 3}.values()
+    oset2 = [3, 9, 27] | oset
+    assertEqual(oset2, [3, 9, 27, 0, 1, 2, 4, 5, 6, 7, 8])
 
-    # Corner case: OrderedDict is not a Sequence, so we don't check for order,
-    # even though it does have the concept of order.
-    assert OrderedSet([1, 2]) != collections.OrderedDict([(2, 2), (3, 1)])
+    # make sure original oset isn't changed
+    assertEqual(oset, lst)
 
+    oset1 = set_t(lst)
+    oset2 = set_t(lst)
 
-def allsame_(iterable, eq=operator.eq):
-    """returns True of all items in iterable equal each other"""
-    iter_ = iter(iterable)
-    try:
-        first = next(iter_)
-    except StopIteration:
-        return True
-    return all(eq(first, item) for item in iter_)
+    oset3 = oset1 | oset2
+    assertEqual(oset3, oset1)
 
+    assertEqual(oset3, oset1.union(oset2))
 
-def check_results_(results, datas, name):
-    """
-    helper for binary operator tests.
+    oset1 |= set_t("abc")
+    assertEqual(oset1, oset2 | "abc")
 
-    check that all results have the same value, but are different items.
-    data and name are used to indicate what sort of tests is run.
-    """
-    if not allsame_(results):
-        raise AssertionError(
-            "Not all same {} for {} with datas={}".format(results, name, datas)
-        )
-    for a, b in it.combinations(results, 2):
-        if not isinstance(a, (bool, int)):
-            assert a is not b, name + " should all be different items"
+    oset1 = set_t(lst)
+    oset1.update("abc")
+    assertEqual(oset1, oset2 | "abc")
 
 
-def _operator_consistency_testdata():
-    """
-    Predefined and random data used to test operator consistency.
-    """
-    # test case 1
-    data1 = OrderedSet([5, 3, 1, 4])
-    data2 = OrderedSet([1, 4])
-    yield data1, data2
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_union_with_iterable(set_t):
+    oset1 = set_t([1])
 
-    # first set is empty
-    data1 = OrderedSet([])
-    data2 = OrderedSet([3, 1, 2])
-    yield data1, data2
+    assertEqual(oset1 | [2, 1], set_t([1, 2]))
+    oset2 = [2] | oset1
+    assertEqual(oset2, set_t([2, 1]))
+    oset2 = [1, 2] | set_t([3, 1, 2, 4])
+    assertEqual(oset2, set_t([1, 2, 3, 4]))
 
-    # second set is empty
-    data1 = OrderedSet([3, 1, 2])
-    data2 = OrderedSet([])
-    yield data1, data2
+    # union with unordered set should work, though the order will be arbitrary
+    oset2 = oset1 | set([2])
+    assertEqual(oset2, set_t([1, 2]))
+    oset2 = set([2]) | oset1
+    assertEqual(oset2, set_t([2, 1]))
 
-    # both sets are empty
-    data1 = OrderedSet([])
-    data2 = OrderedSet([])
-    yield data1, data2
 
-    # random test cases
-    rng = random.Random(0)
-    a, b = 20, 20
-    for _ in range(10):
-        data1 = OrderedSet(rng.randint(0, a) for _ in range(b))
-        data2 = OrderedSet(rng.randint(0, a) for _ in range(b))
-        yield data1, data2
-        yield data2, data1
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_symmetric_difference_with_iterable(set_t):
+    oset1 = set_t([1])
 
+    assertEqual(oset1 ^ [1], set_t([]))
+    assertEqual([1] ^ oset1, set_t([]))
 
-def test_operator_consistency_isect():
-    for data1, data2 in _operator_consistency_testdata():
-        result1 = data1.copy()
-        result1.intersection_update(data2)
-        result2 = data1 & data2
-        result3 = data1.intersection(data2)
-        check_results_([result1, result2, result3], datas=(data1, data2), name="isect")
+    assertEqual(set_t([3, 1, 4, 2]) ^ [3, 4], set_t([1, 2]))
+    assertEqual([3, 1, 4, 2] ^ set_t([3, 4]), set_t([1, 2]))
 
+    assertEqual(set_t([3, 1, 4, 2]) ^ set([3, 4]), set_t([1, 2]))
+    oset2 = set([3, 1, 4]) ^ set_t([3, 4, 2])
+    assertEqual(oset2, set_t([1, 2]))
 
-def test_operator_consistency_difference():
-    for data1, data2 in _operator_consistency_testdata():
-        result1 = data1.copy()
-        result1.difference_update(data2)
-        result2 = data1 - data2
-        result3 = data1.difference(data2)
-        check_results_(
-            [result1, result2, result3], datas=(data1, data2), name="difference"
-        )
 
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_intersection_with_iterable(set_t):
+    assertEqual([1, 2, 3] & set_t([3, 2]), set_t([2, 3]))
+    assertEqual(set_t([3, 2] & set_t([1, 2, 3])), set_t([3, 2]))
 
-def test_operator_consistency_xor():
-    for data1, data2 in _operator_consistency_testdata():
-        result1 = data1.copy()
-        result1.symmetric_difference_update(data2)
-        result2 = data1 ^ data2
-        result3 = data1.symmetric_difference(data2)
-        check_results_([result1, result2, result3], datas=(data1, data2), name="xor")
 
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_difference_with_iterable(set_t):
+    assertEqual(set_t([1, 2, 3, 4]) - [3, 2], set_t([1, 4]))
+    assertEqual([3, 2, 4, 1] - set_t([2, 4]), set_t([3, 1]))
 
-def test_operator_consistency_union():
-    for data1, data2 in _operator_consistency_testdata():
-        result1 = data1.copy()
-        result1.update(data2)
-        result2 = data1 | data2
-        result3 = data1.union(data2)
-        check_results_([result1, result2, result3], datas=(data1, data2), name="union")
 
+@pytest.mark.parametrize("set_t", all_sets)
+def test_isdisjoint(set_t):
+    assertTrue(set_t().isdisjoint(set_t()))
+    assertTrue(set_t([1]).isdisjoint(set_t([2])))
+    assertFalse(set_t([1, 2]).isdisjoint(set_t([2, 3])))
 
-def test_operator_consistency_subset():
-    for data1, data2 in _operator_consistency_testdata():
-        result1 = data1 <= data2
-        result2 = data1.issubset(data2)
-        result3 = set(data1).issubset(set(data2))
-        check_results_([result1, result2, result3], datas=(data1, data2), name="subset")
 
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_index(set_t):
+    oset = set_t("abcd")
+    assertEqual(oset.index("b"), 1)
 
-def test_operator_consistency_superset():
-    for data1, data2 in _operator_consistency_testdata():
-        result1 = data1 >= data2
-        result2 = data1.issuperset(data2)
-        result3 = set(data1).issuperset(set(data2))
-        check_results_(
-            [result1, result2, result3], datas=(data1, data2), name="superset"
-        )
 
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_getitem(set_t):
+    oset = set_t("abcd")
+    assertEqual(oset[2], "c")
 
-def test_operator_consistency_disjoint():
-    for data1, data2 in _operator_consistency_testdata():
-        result1 = data1.isdisjoint(data2)
-        result2 = len(data1.intersection(data2)) == 0
-        check_results_([result1, result2], datas=(data1, data2), name="disjoint")
 
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_getitem_slice(set_t):
+    oset = set_t("abcdef")
+    assertEqual(oset[:2], set_t("ab"))
+    assertEqual(oset[2:], set_t("cdef"))
+    assertEqual(oset[::-1], set_t("fedcba"))
+    assertEqual(oset[1:-1:2], set_t("bd"))
+    assertEqual(oset[1::2], set_t("bdf"))
 
-def test_bitwise_and_consistency():
-    # Specific case that was failing without explicit __and__ definition
-    data1 = OrderedSet([12, 13, 1, 8, 16, 15, 9, 11, 18, 6, 4, 3, 19, 17])
-    data2 = OrderedSet([19, 4, 9, 3, 2, 10, 15, 17, 11, 13, 20, 6, 14, 16, 8])
-    result1 = data1.copy()
-    result1.intersection_update(data2)
-    # This requires a custom & operation apparently
-    result2 = data1 & data2
-    result3 = data1.intersection(data2)
-    check_results_([result1, result2, result3], datas=(data1, data2), name="isect")
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_len(set_t, lst: list):
+    oset = set_t(lst)
+    assertEqual(len(oset), len(lst))
+
+    oset.remove(0)
+    assertEqual(len(oset), len(lst) - 1)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_contains(set_t, lst: list):
+    oset = set_t(lst)
+    assertTrue(1 in oset)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_iter_mutated(set_t, lst: list):
+    oset = set_t(lst)
+    it = iter(oset)
+    oset.add("a")
+
+    with pytest.raises(RuntimeError):
+        next(it)
+
+    it = reversed(oset)
+    oset.add("b")
+
+    with pytest.raises(RuntimeError):
+        next(it)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_iter_and_valid_order(set_t, lst: list):
+    oset = set_t(lst)
+    assertEqual(list(oset), lst)
+
+    oset = set_t(lst + lst)
+    assertEqual(list(oset), lst)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_reverse_order(set_t, lst: list):
+    oset = set_t(lst)
+    assertEqual(list(reversed(oset)), list(reversed(lst)))
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_eq(set_t, lst: list):
+    oset1 = set_t(lst)
+    oset2 = set_t(lst)
+
+    assertNotEqual(oset1, None)
+
+    assertEqual(oset1, oset2)
+    assertEqual(oset1, set(lst))
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", ordered_sets)
+def test_eq_list(set_t, lst: list):
+    assertEqual(set_t(lst), list(lst))
+
+
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_repr(set_t):
+    oset = set_t([1])
+    set_class_name = set_t.__name__
+    assertEqual(repr(oset), f"{set_class_name}([1])")
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_subset(set_t, lst: list):
+    oset1 = set_t([1, 2, 3])
+    oset2 = set_t([1, 2, 3, 4])
+    oset3 = set_t([1, 2, 4, 3])
+    oset4 = set_t([1, 3, 2, 4])
+
+    assertTrue(oset1.isorderedsubset(oset2))
+    assertFalse(oset1.isorderedsubset(oset3))
+    assertFalse(oset2.isorderedsubset(oset3))
+    assertFalse(oset1.isorderedsubset(oset4))
+
+    assertTrue(oset1 < oset2)
+    assertTrue(oset1 < oset3)
+    assertFalse(oset2 < oset3)
+    assertTrue(oset1 < oset4)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_subset_non_consecutive(set_t, lst: list):
+    oset1 = set_t([1, 2, 3])
+    oset2 = set_t([6, 1, 2, 5, 3, 4])
+    oset3 = set_t([6, 1, 2, 5, 4, 3])
+    oset4 = set_t([6, 1, 3, 5, 2, 4])
+
+    assertTrue(oset1.isorderedsubset(oset2, non_consecutive=True))
+    assertTrue(oset1.isorderedsubset(oset3, non_consecutive=True))
+    assertFalse(oset2.isorderedsubset(oset3, non_consecutive=True))
+    assertFalse(oset1.isorderedsubset(oset4, non_consecutive=True))
+
+    assertTrue(oset1 < oset2)
+    assertTrue(oset1 < oset3)
+    assertFalse(oset2 < oset3)
+    assertTrue(oset1 < oset4)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_superset(set_t, lst: list):
+    oset1 = set_t([1, 2, 3])
+    oset2 = set_t([1, 2, 3, 4])
+    oset3 = set_t([1, 2, 4, 3])
+    oset4 = set_t([1, 3, 2, 4])
+
+    assertTrue(oset2.isorderedsuperset(oset1))
+    assertFalse(oset3.isorderedsuperset(oset1))
+    assertFalse(oset3.isorderedsuperset(oset2))
+    assertFalse(oset4.isorderedsuperset(oset1))
+
+    assertTrue(oset2 > oset1)
+    assertTrue(oset3 > oset1)
+    assertFalse(oset3 > oset2)
+    assertTrue(oset4 > oset1)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stable_and_ordered_sets)
+def test_superset_non_consecutive(set_t, lst: list):
+    oset1 = set_t([1, 2, 3])
+    oset2 = set_t([6, 1, 2, 5, 3, 4])
+    oset3 = set_t([6, 1, 2, 5, 4, 3])
+    oset4 = set_t([6, 1, 3, 5, 2, 4])
+
+    assertTrue(oset2.isorderedsuperset(oset1, non_consecutive=True))
+    assertTrue(oset3.isorderedsuperset(oset1, non_consecutive=True))
+    assertFalse(oset3.isorderedsuperset(oset2, non_consecutive=True))
+    assertFalse(oset4.isorderedsuperset(oset1, non_consecutive=True))
+
+    assertTrue(oset2 > oset1)
+    assertTrue(oset3 > oset1)
+    assertFalse(oset3 > oset2)
+    assertTrue(oset4 > oset1)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", all_sets)
+def test_ordering(set_t, lst: list):
+    oset1 = set_t(lst)
+    oset2 = set_t(lst)
+
+    assertLessEqual(oset2, oset1)
+    assertLessEqual(oset2, set(oset1))
+
+    assertGreaterEqual(oset1, oset2)
+    assertGreaterEqual(oset1, set(oset2))
+
+    oset3 = set_t(lst[:-1])
+
+    assertLess(oset3, oset1)
+    assertLess(oset3, set(oset1))
+
+    assertGreater(oset1, oset3)
+    assertGreater(oset1, set(oset3))
+
+    oset4 = set_t(lst[1:])
+
+    assertFalse(oset3 < oset4)
+    assertFalse(oset3 < set(oset4))
+    assertFalse(oset3 >= oset4)
+    assertFalse(oset3 >= set(oset4))
+    assertFalse(oset3 < oset4)
+    assertFalse(oset3 < set(oset4))
+    assertFalse(oset3 >= oset4)
+    assertFalse(oset3 >= set(oset4))
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stableeq_and_ordered_sets)
+def test_ordering_with_lists(set_t, lst: list):
+    oset1 = set_t(lst)
+    oset2 = set_t(lst)
+
+    assertLessEqual(oset2, list(oset1))
+    assertGreaterEqual(oset1, list(oset2))
+
+    oset3 = set_t(lst[:-1])
+
+    assertLess(oset3, list(oset1))
+    assertGreater(oset1, list(oset3))
+
+    oset4 = set_t(lst[1:])
+
+    assertFalse(oset3 < list(oset4))
+    assertFalse(oset3 >= list(oset4))
+    assertFalse(oset3 < list(oset4))
+    assertFalse(oset3 >= list(oset4))
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", ordered_sets)
+def test_eq_reversed_orderedset(set_t, lst: list):
+    oset1 = set_t(lst)
+    oset2 = set_t(reversed(lst))
+    assertNotEqual(oset1, oset2)
+
+
+@pytest.mark.parametrize("lst", datasets)
+@pytest.mark.parametrize("set_t", stableeq_sets)
+def test_eq_reversed_stableset(set_t, lst: list):
+    oset1 = set_t(lst)
+    oset2 = set_t(reversed(lst))
+    assertEqual(oset1, oset2)